Create openvas-nvt-sync script
This is a modified openvas-nvt-sync script, updated with the new nvt source and removing the md5 check as a workaround.  In a future version I will have it add an sha256 check, but for now you have to do that by hand.  Anyway, it gets the script working on debian again.